♻️ refactor(blob): move the test-only FSStore into blob/blobtest
blob.FSStore had no production constructor: NewStoreFromConfig only ever builds
an S3Store, so the filesystem backend survived purely as a test double for the
five packages that need a real, inspectable blob.Store without an S3 server
(internal/authz, internal/agent/session/access x2, internal/server x2).

Keeping it in blob made the package doc and architecture.md claim a filesystem
deployment backend that no config path can select. Moved it to
internal/platform/blob/blobtest (FSStore + NewFSStore + its tests, unchanged
except for calling blob.ValidateKey), leaving blob with exactly one production
implementer of Store. TestNewStoreFromConfig stayed behind in
blob/config_test.go, the only test in the old fs_test.go that tested blob
itself.

Docs fixed in the same commit: internal/platform/blob/doc.go and
web/content/docs/development/architecture.md / .zh.md no longer say "local
filesystem or S3".

Verified: go build ./... , go vet ./internal/... ./cmd/... ./pkg/... ,
go test ./internal/platform/... ./internal/authz/... ./internal/agent/session/...
./internal/server/... ./internal/asset/...
