+++ internal/platform/blob/doc.go
@@ -1,4 +1,5 @@
-// Package blob stores opaque bytes behind one interface, backed either by the
-// local filesystem or by S3-compatible object storage. Callers address content
-// by key and never learn which backend answered.
+// Package blob stores opaque bytes behind one interface, backed by
+// S3-compatible object storage. Callers address content by key and never learn
+// which backend answered. Tests that need a real Store without an S3 server use
+// the filesystem-backed one in blob/blobtest.
 package blob
